implements the version routes
