+++ intavia_backend/main_v2.py
@@ -4,9 +4,9 @@ from fastapi import APIRouter, Depends
 from fastapi_versioning import version, versioned_api_route
 from intavia_backend.models_v1 import PaginatedResponseEntities
 from intavia_backend.query_parameters import Entity_Retrieve
-from .utils import get_query_from_triplestore
+from .utils import flatten_rdf_data, get_query_from_triplestore_v2
 
-router = APIRouter(route_class=versioned_api_route(2))
+router = APIRouter(route_class=versioned_api_route(2, 0))
 
 
 @router.get(
@@ -17,7 +17,7 @@ router = APIRouter(route_class=versioned_api_route(2))
     description="Endpoint that allows to retrive an entity by id.",
 )
 async def retrieve_entity_v2(search: Entity_Retrieve = Depends()):
-    res = get_query_from_triplestore(search, "get_entity_v1.sparql", "search_v3.sparql")
+    res = get_query_from_triplestore_v2(search, "get_entity_v1.sparql")
     pages = math.ceil(int(res[0]["_count"]) / search.limit) if len(res) > 0 else 0
     count = int(res[0]["_count"]) if len(res) > 0 else 0
-    return {"page": search.page, "count": count, "pages": pages, "results": res}
+    return {"page": search.page, "count": count, "pages": pages, "results": flatten_rdf_data(res)}
